+++ .github/workflows/pytester-cov.yml
@@ -28,72 +28,71 @@ name: pytester-cov
 
 on:
   push:
-    branches: [ main ]
+    branches: [main]
   pull_request:
-    branches: [ main ]
+    branches: [main]
 
 jobs:
   build:
-
     runs-on: ubuntu-latest
     env:
       COVERAGE_SINGLE: 80
       COVERAGE_TOTAL: 80
 
     steps:
-    - uses: actions/checkout@v2
-    - name: Set up Python 3.9
-      uses: actions/setup-python@v2
-      with:
-        python-version: 3.9
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install flake8 pytest
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+      - uses: actions/checkout@v2
+      - name: Set up Python 3.9
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.9
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install flake8 pytest
+          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
 
-    - name: pytester-cov
-      id: pytester-cov
-      uses: programmingwithalex/pytester-cov@v1.2.4
-      with:
-        pytest-root-dir: '.'
-        cov-omit-list: 'tests/*, setup.py, src/conftest.py, src/wordle/__init__.py'
-        cov-threshold-single: ${{ env.COVERAGE_SINGLE }}
-        cov-threshold-total: ${{ env.COVERAGE_TOTAL }}
+      - name: pytester-cov
+        id: pytester-cov
+        uses: programmingwithalex/pytester-cov@v1.2.4
+        with:
+          pytest-root-dir: '.'
+          cov-omit-list: 'tests/*, setup.py, src/conftest.py, src/pythle/__init__.py'
+          cov-threshold-single: ${{ env.COVERAGE_SINGLE }}
+          cov-threshold-total: ${{ env.COVERAGE_TOTAL }}
 
-    - name: Coverage single fail - new issue
-      if: ${{ steps.pytester-cov.outputs.cov-threshold-single-fail == 'true' }}
-      uses: nashmaniac/create-issue-action@v1.1
-      with:
-        title: Pytest coverage single falls below minimum ${{ env.COVERAGE_SINGLE }}
-        token: ${{secrets.GITHUB_TOKEN}}
-        assignees: ${{github.actor}}
-        labels: workflow-failed
-        body: ${{ steps.pytester-cov.outputs.output-table }}
+      - name: Coverage single fail - new issue
+        if: ${{ steps.pytester-cov.outputs.cov-threshold-single-fail == 'true' }}
+        uses: nashmaniac/create-issue-action@v1.1
+        with:
+          title: Pytest coverage single falls below minimum ${{ env.COVERAGE_SINGLE }}
+          token: ${{secrets.GITHUB_TOKEN}}
+          assignees: ${{github.actor}}
+          labels: workflow-failed
+          body: ${{ steps.pytester-cov.outputs.output-table }}
 
-    - name: Coverage single fail - exit
-      if: ${{ steps.pytester-cov.outputs.cov-threshold-single-fail == 'true' }}
-      run: |
-        echo "cov single fail ${{ steps.pytester-cov.outputs.cov-threshold-single-fail }}"
-        exit 1
+      - name: Coverage single fail - exit
+        if: ${{ steps.pytester-cov.outputs.cov-threshold-single-fail == 'true' }}
+        run: |
+          echo "cov single fail ${{ steps.pytester-cov.outputs.cov-threshold-single-fail }}"
+          exit 1
 
-    - name: Coverage total fail - new issue
-      if: ${{ steps.pytester-cov.outputs.cov-threshold-total-fail == 'true' }}
-      uses: nashmaniac/create-issue-action@v1.1
-      with:
-        title: Pytest coverage total falls below minimum ${{ env.COVERAGE_TOTAL }}
-        token: ${{secrets.GITHUB_TOKEN}}
-        assignees: ${{github.actor}}
-        labels: workflow-failed
-        body: ${{ steps.pytester-cov.outputs.output-table }}
+      - name: Coverage total fail - new issue
+        if: ${{ steps.pytester-cov.outputs.cov-threshold-total-fail == 'true' }}
+        uses: nashmaniac/create-issue-action@v1.1
+        with:
+          title: Pytest coverage total falls below minimum ${{ env.COVERAGE_TOTAL }}
+          token: ${{secrets.GITHUB_TOKEN}}
+          assignees: ${{github.actor}}
+          labels: workflow-failed
+          body: ${{ steps.pytester-cov.outputs.output-table }}
 
-    - name: Coverage total fail - exit
-      if: ${{ steps.pytester-cov.outputs.cov-threshold-total-fail == 'true' }}
-      run: |
-        echo "cov single fail ${{ steps.pytester-cov.outputs.cov-threshold-total-fail }}"
-        exit 1
+      - name: Coverage total fail - exit
+        if: ${{ steps.pytester-cov.outputs.cov-threshold-total-fail == 'true' }}
+        run: |
+          echo "cov single fail ${{ steps.pytester-cov.outputs.cov-threshold-total-fail }}"
+          exit 1
 
-    - name: Commit pytest coverage table
-      uses: peter-evans/commit-comment@v1
-      with:
-        body: ${{ steps.pytester-cov.outputs.output-table }}
+      - name: Commit pytest coverage table
+        uses: peter-evans/commit-comment@v1
+        with:
+          body: ${{ steps.pytester-cov.outputs.output-table }}
