+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name = wordle
-version = attr: wordle.__version__
+name = pythle
+version = attr: pythle.__version__
 author = Tom Gooding
 description = A terminal-based version of the popular word-guessing game Wordle, built with Python. Created as my first GitHub project! 
 long_description = file: README.md
@@ -25,4 +25,4 @@ where = src
 
 [options.entry_points]
 console_scripts =
-    wordle = wordle.wordle:main
+    pythle = pythle.pythle:main
