+++ src/pythle/__init__.py
@@ -1,0 +1,6 @@
+"""
+A terminal-based version of the popular word-guessing game pythle, built with Python.
+Created as my first GitHub project!
+"""
+
+__version__ = "0.1.0"
