--- src/wordle/__init__.py
@@ -1,6 +1,0 @@
-"""
-A terminal-based version of the popular word-guessing game Wordle, built with Python.
-Created as my first GitHub project!
-"""
-
-__version__ = "0.1.0"
